message_brokers/kafka_example/12_distributed_transactions/requirements.txt
kafka-python
testcontainers
pytest
